Split UndoManager into an orchestrator with per-domain undo strategies (#3403)
UndoManager becomes a pure orchestrator that owns the undo-lock lifecycle and the UndosChanged event, and delegates domain specifics to per-domain IUndoStrategy implementations chosen by the current selection: ElementUndoStrategy (the granular element track) and BehaviorUndoStrategy (the whole-object behavior track). This removes the asymmetry of two inline tracks and sets up a third (animations) as a new strategy rather than another inline region.

Behavior-preserving: both tracks already stored an explicit per-action RedoState (element via the swapped-args reverse diff, behavior via a clone), so standardizing the interface on explicit RedoState required no change to element redo computation -- only relocation. Proven by the audited/backfilled pinning tests, which stay green before and after.

Also: SetStateContentsFrom moved to a shared UndoStateHelper (used by both strategies); CloneWithFixedEnumerations moved onto ElementUndoStrategy; the behavior History-tab wording (GetBehaviorActionDescription) moved from UndosViewModel onto BehaviorUndoStrategy so each domain owns its descriptions. IUndoManager's public surface is unchanged and UndoManager's constructor signature is identical, so DI/composition is untouched.

PR 1 of 2 toward #3399 (animation keyframe undo).
